requirements.txt
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.5
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
efficientnet-pytorch==0.7.1
einops==0.6.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.9.0
fonttools==4.42.1
fsspec==2023.6.0
gmpy2==2.1.2
huggingface-hub==0.16.4
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter-client==7.3.4
jupyter_core==4.12.0
kiwisolver==1.4.5
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
munch==4.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.25.2
opencv-contrib-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.2.1
pretrainedmodels==0.7.4
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
segmentation-models-pytorch==0.3.3
setuptools==68.0.0
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.2.0
timm==0.9.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.1
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
ttach==0.0.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
wheel==0.38.4
